requirements.txt
flask==2.3.3
requests==2.31.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
textblob==0.17.1
nltk==3.8.1
numpy==1.24.3
sqlite3